view and approvel of instructor application

+++ app/Http/Controllers/Admin/SpeakersController.php
@@ -8,21 +8,27 @@ use App\Http\Requests\UpdateSpeakerRequest;
 use App\Models\Event;
 use App\Models\Speaker;
 use App\Services\Event\SpeakerService;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Http\Request;
 
 class SpeakersController extends Controller
 {
+    use AuthorizesRequests;
     public function __construct(protected SpeakerService $speakerService){}
     public function index(){
+        $this->authorize('viewAny', Speaker::class);
         $speakers = $this->speakerService->fetchSpeakers();
         return view("admin.speakers.index", compact('speakers'));
     }
 
     public function create(){
+        $this->authorize('create', Speaker::class);
         return view("admin.speakers.create-speaker");
     }
 
     public function store(CreateSpeakerRequest $request){
+        $this->authorize('create', Speaker::class);
+
         $speaker = $this->speakerService->createSpeaker($request);
         if($speaker){
           return back()->with([
@@ -47,6 +53,8 @@ class SpeakersController extends Controller
     public function update(UpdateSpeakerRequest $request, Speaker $speaker)
     {
         $speaker = $this->speakerService->updateSpeaker($request, $speaker);
+        $this->authorize('update', $speaker);
+
         if($speaker){
             return back()->with([
             "type" => "success",
@@ -59,7 +67,8 @@ class SpeakersController extends Controller
         ]);
     }
 
-    public function destroySpeaker(Speaker $speaker){
+    public function destroySpeaker(Request $request, Speaker $speaker){
+        $this->authorize('delete', $speaker);
         if($this->speakerService->deleteSpeaker($speaker)){
             return back()->with([
                 "type" => "success",
